+++ src/multi_vector_simulator/F0_output.py
@@ -19,6 +19,8 @@ except ModuleNotFoundError:
 from multi_vector_simulator.utils.constants import (
     SIMULATION_SETTINGS,
     PATH_OUTPUT_FOLDER,
+    OUTPUT_FOLDER,
+    LOGFILE,
 )
 from multi_vector_simulator.utils.constants_json_strings import (
     UNIT,
@@ -28,6 +30,10 @@ from multi_vector_simulator.utils.constants_json_strings import (
     RESOURCES,
     KPI_SCALARS_DICT,
     ECONOMIC_DATA,
+    SIMULATION_RESULTS,
+    LOGS,
+    ERRORS,
+    WARNINGS,
 )
 
 r"""
@@ -68,6 +74,13 @@ def evaluate_dict(dict_values, path_pdf_report=None, path_png_figs=None):
 
     logging.info(
         "Summarizing simulation results to results_timeseries and results_scalars_assets."
+    )
+
+    parse_simulation_log(
+        path_log_file=os.path.join(
+            dict_values[SIMULATION_SETTINGS][PATH_OUTPUT_FOLDER], LOGFILE
+        ),
+        dict_values=dict_values,
     )
 
     # storing all flows to exel.
@@ -160,7 +173,6 @@ def store_scalars_to_excel(dict_values):
                 results_scalar_output_file,
                 kpi_set,
             )
-    return
 
 
 def store_timeseries_all_busses_to_excel(dict_values):
@@ -186,7 +198,54 @@ def store_timeseries_all_busses_to_excel(dict_values):
             dict_values[OPTIMIZED_FLOWS][bus].to_excel(open_file, sheet_name=bus)
 
     logging.debug("Saved flows at busses to: %s.", timeseries_output_file)
-    return
+
+
+def parse_simulation_log(path_log_file, dict_values):
+    """Gather a log file with several log messages, this function gathers them all and inputs them into the dict with
+    all input and output parameters up to F0
+
+    Parameters
+    ----------
+    path_log_file: str/None
+        path to the mvs log file
+        Default: None
+
+    dict_values :
+        dict Of all input and output parameters up to F0
+
+    Returns
+    -------
+    Updates the results dictionary with the log messages of the simulation
+
+    """
+    # Dictionaries to gather non-fatal warning and error messages that appear during the simulation
+
+    error_dict, warning_dict = {}, {}
+
+    if path_log_file is None:
+        path_log_file = os.path.join(OUTPUT_FOLDER, LOGFILE)
+
+    with open(path_log_file) as log_messages:
+        log_messages = log_messages.readlines()
+
+    i = j = 0
+
+    # Loop through the list of lines of the log file to check for the relevant log messages and gather them in dicts
+    for line in log_messages:
+        if "ERROR" in line:
+            i = i + 1
+            substrings = line.split(" - ")
+            message_string = substrings[-1]
+            error_dict.update({i: message_string})
+        elif "WARNING" in line:
+            j = j + 1
+            substrings = line.split(" - ")
+            message_string = substrings[-1]
+            warning_dict.update({j: message_string})
+
+    log_dict = {ERRORS: error_dict, WARNINGS: warning_dict}
+
+    dict_values.update({SIMULATION_RESULTS: {LOGS: log_dict}})
 
 
 def store_as_json(dict_values, output_folder=None, file_name=None):